Build breaks if PACKAGECONFIG excludes examples

+++ recipes-graphics/libegt/libegt_1.7.bb
@@ -73,8 +73,14 @@ INSANE_SKIP:${PN} = "dev-so"
 #deleted audio files to avoid check_data_file_clashes error
 do_install:append() {
     rm -f ${D}/usr/lib/libegt.a
-    rm -f ${D}/usr/share/egt/examples/audioplayer/*.mp3
-    rm -f ${D}/usr/share/egt/examples/drummachine/*.wav
+
+    # Only attempt file removal if examples are built
+    if ${@bb.utils.contains('PACKAGECONFIG', 'examples', 'true', 'false', d)}; then
+        rm -f ${D}/usr/share/egt/examples/audioplayer/*.mp3
+        rm -f ${D}/usr/share/egt/examples/drummachine/*.wav
+        sed -i -e '2,/<\/entry>/d' ${D}/usr/share/egt/examples/video/video.xml
+    fi
+
     sed -e 's@[^ ]*-ffile-prefix-map=[^ "]*@@g' \
         -e 's@[^ ]*-fdebug-prefix-map=[^ "]*@@g' \
         -e 's@[^ ]*-fmacro-prefix-map=[^ "]*@@g' \
